DP side: * Fixed issues with char_shortcuts * More table cleanup and overhead reduction
git-svn-id: http://svn.l2jfree.com/l2jfree@5856 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ trunk/l2jfree-datapack/sql/character_shortcuts.sql
@@ -7,7 +7,7 @@ CREATE TABLE IF NOT EXISTS `character_shortcuts` (
   `page` TINYINT NOT NULL,
   `type` TINYINT NOT NULL,
   `shortcut_id` INT NOT NULL,
-  `level` TINYINT NOT NULL,
+  `level` SMALLINT(3) NOT NULL,
   `class_index` TINYINT UNSIGNED NOT NULL,
   PRIMARY KEY (`charId`,`slot`,`page`,`class_index`),
   KEY `shortcut_id` (`shortcut_id`)
